format, refactor, and organize
