+++ CMakeLists.txt
@@ -6,10 +6,9 @@ set(CMAKE_CXX_VISIBILITY_PRESET hidden)
 
 project(OptimizedSongSize VERSION 1.0.0)
 
-add_library(${PROJECT_NAME} SHARED
-    src/main.cpp
-    # Add your cpp files here
-)
+file(GLOB_RECURSE SOURCES src/*.cpp src/*.hpp)
+
+add_library(${PROJECT_NAME} SHARED ${SOURCES})
 
 if (NOT DEFINED ENV{GEODE_SDK})
     message(FATAL_ERROR "Unable to find Geode SDK! Please define GEODE_SDK environment variable to point to Geode")
